add test for clicking create room on Dashboard
