+++ server/src/utils/getRandomItem.js
@@ -1,0 +1,4 @@
+export const getRandomItem = list => {
+  var i = Math.floor(Math.random() * list.length);
+  return list[i];
+};
