+++ server/src/utils/nouns.js
@@ -1,0 +1,101 @@
+const singularNouns = [
+  'ninja',
+  'chair',
+  'pancake',
+  'statue',
+  'unicorn',
+  'laser',
+  'senor',
+  'bunny',
+  'captain',
+  'cupcake',
+  'carrot',
+  'glitter',
+  'potato',
+  'salad',
+  'toejam',
+  'toilet',
+  'exorcism',
+  'ice cream',
+  'ukulele',
+  'kazoo',
+  'banjo',
+  'opera singer',
+  'trampoline',
+  'carousel',
+  'carnival',
+  'locomotive',
+  'hot air balloon',
+  'animator',
+  'artisan',
+  'artist',
+  'colorist',
+  'inker',
+  'coppersmith',
+  'director',
+  'designer',
+  'flatter',
+  'stylist',
+  'leadman',
+  'limner',
+  'make-up artist',
+  'model',
+  'musician',
+  'penciller',
+  'producer',
+  'scenographer',
+  'set decorator',
+  'silversmith',
+  'teacher',
+  'auto mechanic',
+  'beader',
+  'bobbin boy',
+  'clerk of the chapel',
+  'filling station attendant',
+  'foreman',
+  'maintenance engineering',
+  'mechanic',
+  'miller',
+  'moldmaker',
+  'panel beater',
+  'patternmaker',
+  'plant operator',
+  'plumber',
+  'sawfiler',
+  'shop foreman',
+  'soaper',
+  'stationary engineer',
+  'wheelwright',
+]
+
+const pluralNouns = [
+  'rainbows',
+  'nibblets',
+  'gnomes',
+  'curtains',
+  'beets',
+  'stick figures',
+  'mermaid eggs',
+  'sea barnacles',
+  'dragons',
+  'jellybeans',
+  'snakes',
+  'dolls',
+  'bushes',
+  'cookies',
+  'apples',
+  'circus',
+  'praying mantis',
+  'woodworkers',
+];
+
+const nouns = [
+  ...singularNouns,
+  ...pluralNouns
+];
+
+export {
+  singularNouns,
+  pluralNouns,
+  nouns
+};
