+++ src/graphql/addRoom.js
@@ -1,7 +1,7 @@
 import { gql } from 'apollo-boost';
 
 const ADD_ROOM = gql`
-  mutation AddRoom($name: String!) {
+  mutation AddRoom($name: String) {
     addRoom(name: $name) {
       id
       name
